+++ .github/workflows/deploycfrc.yml
@@ -36,7 +36,7 @@ jobs:
           path: |
             ~/.npm
             ./.cache
-            ./buildtools/node_modules
+            ./tools/node_modules
           key: ${{ runner.os }}-bunny-${{ hashFiles('**/.cache', '**/package-lock.json', '**/manifest.json') }}
           restore-keys: ${{ runner.os }}-bunny-
 
@@ -47,11 +47,11 @@ jobs:
           check-latest: true
 
       - name: "Download NPM packages"
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npm ci
 
       - name: "Check environmental variables"
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npx gulp check
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
@@ -60,24 +60,24 @@ jobs:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: "Build everything"
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npx gulp buildAll
         env:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: "Prune cache"
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npx gulp pruneCache
         env:
           CFCORE_API_TOKEN: ${{ secrets.CFCORE_API_TOKEN }}
 
       - name: "Zip everything"
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npx gulp zipAll
 
       - name: "Deploy to CurseForge"
         env:
           CURSEFORGE_PROJECT_ID: ${{ secrets.CURSEFORGE_PROJECT_ID }}
           CURSEFORGE_API_TOKEN: ${{ secrets.CURSEFORGE_API_TOKEN }}
-        working-directory: ./buildtools
+        working-directory: ./tools
         run: npx gulp deployCurseForgeBeta
