--- buildtools/tasks/shared/transforms/index.ts
@@ -1,5 +1,0 @@
-import quest from "./quest_i18n";
-import scannable from "./scannable";
-import version from "./version";
-
-export default [quest, scannable, version];
