+++ tools/buildConfig.ts
