+++ tools/gulpfile.ts
@@ -3,6 +3,9 @@ import * as gulp from "gulp";
 import pruneCacheTask from "./tasks/misc/pruneCache";
 export const pruneCache = pruneCacheTask;
 
+import * as quest from "./tasks/github/quest";
+export const transformQB = quest.transformQuestBook;
+
 import sharedTasks from "./tasks/shared";
 import clientTasks from "./tasks/client";
 import serverTasks from "./tasks/server";
