+++ tools/tasks/deploy/curseforge.ts
