+++ tools/tasks/misc/pruneCache.ts
