+++ tools/tasks/shared/transforms/index.ts
@@ -1,0 +1,4 @@
+import scannable from "./scannable";
+import version from "./version";
+
+export default [scannable, version];
