+++ tools/tasks/shared/transforms/scannable.ts
