+++ tools/types/curseForge.ts
