+++ tools/types/hashDef.ts
