+++ tools/types/mainMenuConfig.ts
