+++ tools/types/modpackManifest.ts
