+++ tools/util/curseForgeAPI.ts
