+++ tools/util/hashes.ts
