+++ tools/util/util.ts
